+++ application/controllers/Sucursales.php
@@ -5,9 +5,15 @@ class Sucursales extends CI_Controller {
 	public function __construct(){
 		parent::__construct();
 		$this->load->model("Sucursales_model");
+		$this->load->library(['ion_auth', 'form_validation']);
 	}
 
 	public function index(){
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
+
 		$this->load->view("Sucursales/index");
 	}
 
@@ -65,6 +71,11 @@ class Sucursales extends CI_Controller {
 
 	public function ajax_add()
 	{
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
+
 		$this->_validate();		
 		$data = array(
 				'nombre' => $this->input->post('nombre'),
@@ -79,12 +90,22 @@ class Sucursales extends CI_Controller {
 
 	public function ajax_edit($id)
 	{
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
+
 		$data = $this->Sucursales_model->get_by_id($id);
 		echo json_encode($data);
 	}
 
 	public function ajax_update()
 	{
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
+
 		$this->_validate();
 		$data = array(
 				'nombre' => $this->input->post('nombre'),
@@ -99,12 +120,22 @@ class Sucursales extends CI_Controller {
 
 	public function ajax_delete($id)
 	{	
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
+
 		$this->Sucursales_model->delete_by_id($id);
 		echo json_encode(array("status" => TRUE));
 	}
 
 	public function ajax_enabled($id)
 	{		
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
+
 		$this->Sucursales_model->enabled_by_id($id);
 		echo json_encode(array("status" => TRUE));
 	}
@@ -131,6 +162,11 @@ class Sucursales extends CI_Controller {
 	}
 
 	public function createXLS() {
+
+		if (!$this->ion_auth->logged_in() || !$this->ion_auth->is_admin())
+		{
+			redirect('auth', 'refresh');
+		}
 
        $this->load->library('excel');
        $empInfo = $this->Sucursales_model->get_all_export();
